worked on animal bonus exercise, created animal bonus package

+++ src/shapes/Rectangle.java
@@ -2,20 +2,20 @@ package shapes;
 
 public class Rectangle {
 
-    protected int length;
-    protected int width;
+    protected double length;
+    protected double width;
 
-    public Rectangle(int length, int width){
+    public Rectangle(double length, double width){
         this.length = length;
         this.width = width;
     }
 
-    public int getArea(){
+    public double getArea(){
         System.out.println("Using Rectangle class!");
         return length*width;
     }
 
-    public int getPerimeter(){
+    public double getPerimeter(){
         System.out.println("Using Rectangle class!");
         return length*2 + width*2;
     }
